+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # gcf-helper CHANGELOG
 
+## 2019-03-21, Version 0.4.0
+
+* added lazy loading of pubsub client
+* added lazy loading of bigquery client
+* added `writeBigQueryRows` method
+* added `getPubSubDataFromEvent` method
+* added `setup` export for ease of use
+* added unit test
+* added documentation to describe helping options
+
 ## 2019-03-19, Version 0.2.0-0.3.0
 
 * reading default FunctionOptions from environment variables (see DOC.md)